Rename diagnostic UI
