+++ diag/app/package.json
