+++ diag/app/src/App.tsx
