+++ diag/app/src/Instance.tsx
