+++ diag/app/src/client.ts
