+++ diag/app/src/index.tsx
