+++ diag/app/src/react-app-env.d.ts
