+++ diag/app/tsconfig.json
